ref: Splitting api/event_search.py up into 3 files
